Threading fixes
