+++ Updater/source/path.py
@@ -17,7 +17,7 @@ import logging
 import os
 import pathlib
 import sys
-from source.portablemc.standard import Context
+from portablemc.standard import Context
 
 logger = logging.getLogger(__name__)
 
@@ -41,6 +41,9 @@ STORE_DIR = pathlib.Path(STORE_DIR)
 MAIN_DIR = pathlib.Path(os.path.join(STORE_DIR, "minecraft"))
 WORK_DIR = pathlib.Path(os.path.join(STORE_DIR, "mcdata"))
 CONTEXT = Context(MAIN_DIR, WORK_DIR)
+GLOBAL_KILL = pathlib.Path(os.path.join(STORE_DIR, "GLOBAL_KILL"))
+if GLOBAL_KILL.exists():
+    GLOBAL_KILL.unlink()
 
 os.makedirs(STORE_DIR, exist_ok=True)
 os.makedirs(MAIN_DIR, exist_ok=True)
