--- Updater/source/portablemc/__init__.py
@@ -1,13 +1,0 @@
-"""Main module for PortableMC API.
-
-Please read the API documentation for more general information about the architecture of
-the API and examples codes. You should also read the API documentation about 
-*Versioning and stability* to understand which classes you should consider part of the
-public API, and therefore stable across minor and patch version bumps.
-"""
-
-LAUNCHER_NAME = "portablemc"
-LAUNCHER_VERSION = "4.3.0"
-LAUNCHER_AUTHORS = ["Théo Rozier <contact@theorozier.fr>", "Contributors"]
-LAUNCHER_COPYRIGHT = "PortableMC  Copyright (C) 2021-2024  Théo Rozier"
-LAUNCHER_URL = "https://github.com/mindstorm38/portablemc"
